Label benchmark capacity models
Record nominal capacity and whether each channel uses a shared bound or per-ring HWM. Keep legacy benchmark rows readable.

Split chart parsing and rendering into focused modules and regenerate both tracked charts with explicit capacity notes.
