+++ benches/mpmc.rs
@@ -60,7 +60,8 @@ struct Row {
     producers: usize,
     consumers: usize,
     capacity_per_sender: usize,
-    total_capacity: usize,
+    nominal_capacity: usize,
+    capacity_model: &'static str,
     seconds: f64,
     items: u64,
     items_per_sec: f64,
@@ -454,7 +455,12 @@ where
         producers: config.producers,
         consumers: config.consumers,
         capacity_per_sender: config.capacity_per_sender,
-        total_capacity: config.total_capacity(),
+        nominal_capacity: config.total_capacity(),
+        capacity_model: if implementation == "fanring-mpmc" {
+            "per-ring-hwm-with-staging"
+        } else {
+            "shared-bound"
+        },
         seconds: elapsed.as_secs_f64(),
         items: received,
         items_per_sec: received as f64 / elapsed.as_secs_f64(),
